+++ frontend/form-filter-things.php
@@ -47,7 +47,7 @@ function filter_things( $args, $options ) {
 		return;
 	}
 
-	$post_type_name = format_list_of_things( $args['post_type'], 'and' );
+	$post_type_names = format_list_of_things( get_post_type_names( $args['post_type'] ), 'and' );
 	?>
 	<div
 		class="thing-filters__button__container"
@@ -59,7 +59,7 @@ function filter_things( $args, $options ) {
 					// Translators: %s is the post type name.
 					__( 'Filter %s', 'list-things' )
 				),
-				esc_html( strtolower( $post_type_name ) )
+				esc_html( strtolower( $post_type_names ) )
 			);
 			?>
 		</p>
@@ -67,50 +67,62 @@ function filter_things( $args, $options ) {
 			<?php esc_html_e( 'Show filters', 'list-things' ); ?>
 		</button>
 	</div>
-	<form
-		id="thing-filters-<?php echo esc_attr( $options['things_section_id'] ); ?>"
-		class="thing-filters__form row gap-xs wrap" 
-		role="search"
-		onsubmit="return false;"
-	>
-		<?php
-		foreach ( $taxonomies as $taxonomy ) {
-			$tax_obj = get_taxonomy( $taxonomy );
-			if (
-				'all' !== $options['filters']
-				&& ! in_array( $taxonomy, $options['filters'], true )
-			) {
-				continue;
-			}
-			?>
-			<fieldset class="thing-filter <?php echo esc_attr( $tax_obj->name ); ?>-filter">
-				<legend class="thing-filter__legend"><?php echo esc_html( $tax_obj->labels->name ); ?></legend>
-				<?php
-					$terms = get_terms(
-						array(
-							'hide_empty' => true,
-							'taxonomy'   => $tax_obj->name,
-						)
-					);
-
-				foreach ( $terms as $term ) {
-					?>
-						<div class="thing-filter__term row gap-xxs">
-							<input
-								id="<?php echo esc_attr( $term->name ); ?>"
-								type="checkbox"
-							>
-							<label for="<?php echo esc_attr( $term->name ); ?>">
-								<?php echo esc_html( $term->name . ' (' . $term->count . ')' ); ?>
-							</label>
-						</div>
-						<?php
+	<div class="thing-filters__form__container">
+		<p class="list-things-label">
+			<strong>
+				<?php esc_html_e( 'Filters', 'list-things' ); ?>
+			</strong>
+			<button class="reset-thing-filters__button wp-element-button list-things-button-link">
+				<?php esc_html_e( 'Reset filters', 'list-things' ); ?>
+			</button>
+		</p>
+		<form
+			id="<?php echo esc_attr( $options['things_section_id'] ); ?>__thing-filters__form"
+			class="thing-filters__form row gap-xs wrap" 
+			role="search"
+			onsubmit="return false;"
+		>
+			<?php
+			foreach ( $taxonomies as $taxonomy ) {
+				$tax_obj = get_taxonomy( $taxonomy );
+				if (
+					'all' !== $options['filters']
+					&& ! in_array( $taxonomy, $options['filters'], true )
+				) {
+					continue;
 				}
 				?>
-			</fieldset>
-			<?php
-		}
-		?>
-	</form>
+				<fieldset class="thing-filter <?php echo esc_attr( $tax_obj->name ); ?>-filter" data-taxonomy="<?php echo esc_attr( $tax_obj->name ); ?>">
+					<legend class="thing-filter__legend"><?php echo esc_html( $tax_obj->labels->name ); ?></legend>
+					<?php
+						$terms = get_terms(
+							array(
+								'hide_empty' => true,
+								'taxonomy'   => $tax_obj->name,
+							)
+						);
+	
+					foreach ( $terms as $term ) {
+						$term_id = $options['things_section_id'] . '__' . $tax_obj->name . '__' . $term->name;
+						?>
+							<div class="thing-filter__term row gap-xxs">
+								<input
+									id="<?php echo esc_attr( $term_id ); ?>"
+									type="checkbox"
+									data-term="<?php echo esc_attr( $term->name ); ?>"
+								>
+								<label for="<?php echo esc_attr( $term_id ); ?>">
+									<?php echo esc_html( $term->name . ' (' . $term->count . ')' ); ?>
+								</label>
+							</div>
+							<?php
+					}
+					?>
+				</fieldset>
+				<?php
+			}
+			?>
+		</form>
+	</div>
 	<?php
 }