+++ frontend/shortcodes.php
@@ -7,7 +7,7 @@
  * @file       shortcodes.php
  * @package    list-things
  * @subpackage frontend
- * @since      1.0.0
+ * @since      0.1.0
  */
 
 namespace List_Things;
@@ -30,6 +30,7 @@ function list_things_shortcode( $atts ) {
 	 * - post__not_in
 	 * - sort_buttons
 	 * - classes
+	 * - filters
 	 */
 	$atts['post_status']  = preg_split( '/, */', $atts['post_status'], -1, PREG_SPLIT_NO_EMPTY );
 	$atts['post_type']    = preg_split( '/, */', $atts['post_type'], -1, PREG_SPLIT_NO_EMPTY );
@@ -39,11 +40,13 @@ function list_things_shortcode( $atts ) {
 		$atts['sort_buttons'] = $atts['sort_buttons'] ? preg_split( '/, */', $atts['sort_buttons'], -1, PREG_SPLIT_NO_EMPTY ) : '';
 	}
 	$atts['classes'] = $atts['classes'] ? preg_split( '/, */', $atts['classes'], -1, PREG_SPLIT_NO_EMPTY ) : '';
-	$atts            = format_vals( $atts );
-
-	// echo '<pre>';
-	// var_dump($atts);
-	// echo '</pre>';
+	if (
+		$atts['filters']
+		&& 'all' !== trim( $atts['filters'] )
+	) {
+		$atts['filters'] = $atts['filters'] ? preg_split( '/, */', $atts['filters'], -1, PREG_SPLIT_NO_EMPTY ) : '';
+	}
+	$atts = format_vals( $atts );
 
 	$args = array(
 		'order'          => $atts['order'],
@@ -64,29 +67,23 @@ function list_things_shortcode( $atts ) {
 		}
 	}
 
-	// echo '<pre>';
-	// var_dump($args);
-	// echo '</pre>';
-
 	$options = array(
 		'grid_cols'      => $atts['grid_cols'],
 		'hide_title'     => $atts['hide_title'],
-		'layout'         => $atts['layout'],
 		'show_excerpt'   => $atts['show_excerpt'],
+		'layout'         => $atts['layout'],
 		'paginate'       => $atts['paginate'],
 		'show_read_more' => $atts['show_read_more'],
 		'show_search'    => $atts['show_search'],
 		'show_sort'      => $atts['show_sort'],
+		'show_filters'   => $atts['show_filters'],
+		'filters'        => $atts['filters'],
 		'sort_buttons'   => $atts['sort_buttons'],
 		'spacing'        => $atts['spacing'],
 		'show_thumbnail' => $atts['show_thumbnail'],
 		'title_tag'      => $atts['title_tag'],
 		'classes'        => $atts['classes'],
 	);
-
-	// echo '<pre>';
-	// var_dump($options);
-	// echo '</pre>';
 
 	wp_enqueue_style( 'list-things', LIST_THINGS_DIR_URL . 'assets/css/thing-styles.css', array(), LIST_THINGS_VERSION );
 
@@ -98,6 +95,10 @@ function list_things_shortcode( $atts ) {
 	wp_register_script( 'sort-things', LIST_THINGS_DIR_URL . 'assets/js/sort-things.js', array( 'jquery' ), LIST_THINGS_VERSION, true );
 	wp_localize_script( 'sort-things', 'vars', array( 'ajaxurl' => admin_url( 'admin-ajax.php' ) ) );
 	wp_enqueue_script( 'sort-things' );
+
+	wp_register_script( 'filter-things', LIST_THINGS_DIR_URL . 'assets/js/filter-things.js', array( 'jquery' ), LIST_THINGS_VERSION, true );
+	wp_localize_script( 'filter-things', 'vars', array( 'ajaxurl' => admin_url( 'admin-ajax.php' ) ) );
+	wp_enqueue_script( 'filter-things' );
 
 	wp_register_script( 'paginate-things', LIST_THINGS_DIR_URL . 'assets/js/paginate-things.js', array( 'jquery' ), LIST_THINGS_VERSION, true );
 	wp_localize_script( 'paginate-things', 'vars', array( 'ajaxurl' => admin_url( 'admin-ajax.php' ) ) );